removed not needed request builder; implemented transfer to saving goals

+++ src/main/java/com/starlingbank/roundUpSaving/model/Amount.java
@@ -1,4 +1,4 @@
 package com.starlingbank.roundUpSaving.model;
 
-public record Amount(String currency, long minorUnits ) {
+public record Amount(String currency, int minorUnits ) {
 }
